requirements.txt
anyio==4.12.1
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
Cython==3.2.4
exceptiongroup==1.3.1
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
fonttools==4.61.1
fsspec==2026.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==26.0
paddleocr==2.8.0
paddlepaddle==3.2.2
pandas==2.3.3
Pillow==10.1.0
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.3
rapidfuzz==3.6.1
requests==2.32.5
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.2
six==1.17.0
soupsieve==2.8.3
sympy==1.14.0
termcolor==3.3.0
tifffile==2025.5.10
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.2
triton==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.2.89
ultralytics-thop==2.0.18
urllib3==2.6.3
Werkzeug==3.1.5